+++ homeworks/nataliia.koval_NataliyaKoval/4-modules/index.html
@@ -37,6 +37,7 @@
         </div>
         
     </div>
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
+    
 </body>
 </html>
